User commit

+++ src/main/java/com/example/demo/entity/Student.java
@@ -2,14 +2,14 @@ package com.example.demo.entity;
 
 import java.time.LocalDate;
 public class Student {
-    private int id;
+    private Long id;
     private String name;
     private String email;
     
     public int getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
@@ -25,7 +25,7 @@ public class Student {
         this.email = email;
     }
     
-    public Student(int id, String name, String email, LocalDate dob, float cgpa) {
+    public Student(Long id, String name, String email, LocalDate dob, float cgpa) {
         this.id = id;
         this.name = name;
         this.email = email;
